Fix Auth bugs
There was a bug that was not allowing  the auth0 callback to work correctly. belive it was an issue with the audience variable being used as an env variable. Changed from the setup file to a .env file.

Also logout endpoint was not working. Was using the wrong logout uri in auth0 app settings.

Function for check permissions in auth file not correct format. Fixed and permission is working now.